+++ src/js/build.js
@@ -10,12 +10,12 @@
       var cropBox;
       var face;
 
-      if (!this.ready) {
+      if (!this.isLoaded) {
         return;
       }
 
       // Unbuild first when replace
-      if (this.built) {
+      if (this.isBuilt) {
         this.unbuild();
       }
 
@@ -45,11 +45,11 @@
       this.initPreview();
       this.bind();
 
-      // Format aspect ratio (0 -> NaN)
       options.aspectRatio = max(0, options.aspectRatio) || NaN;
+      options.viewMode = max(0, min(3, round(options.viewMode))) || 0;
 
       if (options.autoCrop) {
-        this.cropped = true;
+        this.isCropped = true;
 
         if (options.modal) {
           addClass(dragBox, CLASS_MODAL);
@@ -84,14 +84,9 @@
         addClass(querySelectorAll(cropBox, '.cropper-point'), CLASS_HIDDEN);
       }
 
-      this.setDragMode(
-        options.dragCrop ? ACTION_CROP :
-        options.movable ? ACTION_MOVE :
-        ACTION_NONE
-      );
-
+      this.setDragMode(options.dragMode);
       this.render();
-      this.built = true;
+      this.isBuilt = true;
       this.setData(options.data);
 
       // Call the built asynchronously to keep "image.cropper" is defined
@@ -104,16 +99,16 @@
           options.crop.call(element, this.getData());
         }
 
-        this.complete = true;
+        this.isCompleted = true;
       }, this), 0);
     },
 
     unbuild: function () {
-      if (!this.built) {
+      if (!this.isBuilt) {
         return;
       }
 
-      this.built = false;
+      this.isBuilt = false;
       this.initialImageData = null;
 
       // Clear `initialCanvasData` is necessary when replace