+++ src/js/preview.js
@@ -38,7 +38,7 @@
           'display:block;width:100%;height:auto;' +
           'min-width:0!important;min-height:0!important;' +
           'max-width:none!important;max-height:none!important;' +
-          'image-orientation:0deg!important"'
+          'image-orientation:0deg!important;"'
         );
 
         empty(element);
@@ -48,7 +48,11 @@
 
     resetPreview: function () {
       each(this.previews, function (element) {
-        element.innerHTML = getData(element, DATA_PREVIEW).html;
+        var data = getData(element, DATA_PREVIEW);
+
+        element.style.width = data.width + 'px';
+        element.style.height = data.height + 'px';
+        element.innerHTML = data.html;
         removeData(element, DATA_PREVIEW);
       });
     },
@@ -64,7 +68,7 @@
       var left = cropBoxData.left - canvasData.left - imageData.left;
       var top = cropBoxData.top - canvasData.top - imageData.top;
 
-      if (!this.cropped || this.disabled) {
+      if (!this.isCropped || this.isDisabled) {
         return;
       }
 
@@ -73,18 +77,17 @@
         'height:' + height + 'px;' +
         'margin-left:' + -left + 'px;' +
         'margin-top:' + -top + 'px;' +
-        'transform:' + getTransform(imageData)
+        'transform:' + getTransform(imageData) + ';'
       );
 
       each(this.previews, function (element) {
+        var imageStyle = querySelector(element, 'img').style;
         var data = getData(element, DATA_PREVIEW);
         var originalWidth = data.width;
         var originalHeight = data.height;
         var newWidth = originalWidth;
         var newHeight = originalHeight;
         var ratio = 1;
-        var elementStyle = element.style;
-        var imageStyle = querySelector(element, 'img').style;
 
         if (cropBoxWidth) {
           ratio = originalWidth / cropBoxWidth;
@@ -97,8 +100,8 @@
           newHeight = originalHeight;
         }
 
-        elementStyle.width = newWidth + 'px';
-        elementStyle.height = newHeight + 'px';
+        element.style.width = newWidth + 'px';
+        element.style.height = newHeight + 'px';
         imageStyle.width = width * ratio + 'px';
         imageStyle.height = height * ratio + 'px';
         imageStyle.marginLeft = -left * ratio + 'px';