+++ src/js/render.js
@@ -6,7 +6,7 @@
 
       this.renderCanvas();
 
-      if (this.cropped) {
+      if (this.isCropped) {
         this.renderCropBox();
       }
     },
@@ -28,7 +28,7 @@
 
       cropper.style.cssText = (
         'width:' + containerData.width + 'px;' +
-        'height:' + containerData.height + 'px'
+        'height:' + containerData.height + 'px;'
       );
 
       addClass(element, CLASS_HIDDEN);
@@ -37,6 +37,7 @@
 
     // Canvas (image wrapper)
     initCanvas: function () {
+      var viewMode = this.options.viewMode;
       var containerData = this.containerData;
       var containerWidth = containerData.width;
       var containerHeight = containerData.height;
@@ -51,57 +52,74 @@
           };
 
       if (containerHeight * aspectRatio > containerWidth) {
-        canvasData.height = containerWidth / aspectRatio;
+        if (viewMode === 3) {
+          canvasData.width = containerHeight * aspectRatio;
+        } else {
+          canvasData.height = containerWidth / aspectRatio;
+        }
       } else {
-        canvasData.width = containerHeight * aspectRatio;
+        if (viewMode === 3) {
+          canvasData.height = containerWidth / aspectRatio;
+        } else {
+          canvasData.width = containerHeight * aspectRatio;
+        }
       }
 
       canvasData.oldLeft = canvasData.left = (containerWidth - canvasData.width) / 2;
       canvasData.oldTop = canvasData.top = (containerHeight - canvasData.height) / 2;
 
       this.canvasData = canvasData;
+      this.isLimited = (viewMode === 1 || viewMode === 2);
       this.limitCanvas(true, true);
       this.initialImageData = extend({}, imageData);
       this.initialCanvasData = extend({}, canvasData);
     },
 
-    limitCanvas: function (size, position) {
+    limitCanvas: function (isSizeLimited, isPositionLimited) {
       var options = this.options;
-      var strict = options.strict;
+      var viewMode = options.viewMode;
       var containerData = this.containerData;
       var containerWidth = containerData.width;
       var containerHeight = containerData.height;
       var canvasData = this.canvasData;
       var aspectRatio = canvasData.aspectRatio;
       var cropBoxData = this.cropBoxData;
-      var cropped = this.cropped && cropBoxData;
-      var initialCanvasData = this.initialCanvasData || canvasData;
+      var isCropped = this.isCropped && cropBoxData;
       var minCanvasWidth;
       var minCanvasHeight;
+      var newCanvasLeft;
+      var newCanvasTop;
 
-      if (size) {
+      if (isSizeLimited) {
         minCanvasWidth = num(options.minCanvasWidth) || 0;
         minCanvasHeight = num(options.minCanvasHeight) || 0;
 
-        if (strict) {
-          if (minCanvasWidth) {
-            minCanvasWidth = max(
-              minCanvasWidth,
-              cropped ? cropBoxData.width : initialCanvasData.width
-            );
-          } else if (minCanvasHeight) {
-            minCanvasHeight = max(
-              minCanvasHeight,
-              cropped ? cropBoxData.height : initialCanvasData.height
-            );
-          } else if (cropped) {
-            minCanvasWidth = cropBoxData.width;
-            minCanvasHeight = cropBoxData.height;
+        if (viewMode) {
+          if (viewMode > 1) {
+            minCanvasWidth = max(minCanvasWidth, containerWidth);
+            minCanvasHeight = max(minCanvasHeight, containerHeight);
 
-            if (minCanvasHeight * aspectRatio > minCanvasWidth) {
-              minCanvasWidth = minCanvasHeight * aspectRatio;
-            } else {
-              minCanvasHeight = minCanvasWidth / aspectRatio;
+            if (viewMode === 3) {
+              if (minCanvasHeight * aspectRatio > minCanvasWidth) {
+                minCanvasWidth = minCanvasHeight * aspectRatio;
+              } else {
+                minCanvasHeight = minCanvasWidth / aspectRatio;
+              }
+            }
+          } else {
+            if (minCanvasWidth) {
+              minCanvasWidth = max(minCanvasWidth, isCropped ? cropBoxData.width : 0);
+            } else if (minCanvasHeight) {
+              minCanvasHeight = max(minCanvasHeight, isCropped ? cropBoxData.height : 0);
+            } else if (isCropped) {
+              minCanvasWidth = cropBoxData.width;
+              minCanvasHeight = cropBoxData.height;
+
+              if (minCanvasHeight * aspectRatio > minCanvasWidth) {
+                minCanvasWidth = minCanvasHeight * aspectRatio;
+              } else {
+                minCanvasHeight = minCanvasWidth / aspectRatio;
+              }
             }
           }
         }
@@ -124,16 +142,40 @@
         canvasData.maxHeight = Infinity;
       }
 
-      if (position) {
-        if (strict) {
-          canvasData.minLeft = cropped ?
-            min(cropBoxData.left, (cropBoxData.left + cropBoxData.width) - canvasData.width) :
-            min(0, containerWidth - canvasData.width);
-          canvasData.minTop = cropped ?
-            min(cropBoxData.top, (cropBoxData.top + cropBoxData.height) - canvasData.height) :
-            min(0, containerHeight - canvasData.height);
-          canvasData.maxLeft = cropped ? cropBoxData.left : max(0, containerWidth - canvasData.width);
-          canvasData.maxTop = cropped ? cropBoxData.top : max(0, containerHeight - canvasData.height);
+      if (isPositionLimited) {
+        if (viewMode) {
+          newCanvasLeft = containerWidth - canvasData.width;
+          newCanvasTop = containerHeight - canvasData.height;
+
+          canvasData.minLeft = min(0, newCanvasLeft);
+          canvasData.minTop = min(0, newCanvasTop);
+          canvasData.maxLeft = max(0, newCanvasLeft);
+          canvasData.maxTop = max(0, newCanvasTop);
+
+          if (isCropped && this.isLimited) {
+            canvasData.minLeft = min(
+              cropBoxData.left,
+              cropBoxData.left + cropBoxData.width - canvasData.width
+            );
+            canvasData.minTop = min(
+              cropBoxData.top,
+              cropBoxData.top + cropBoxData.height - canvasData.height
+            );
+            canvasData.maxLeft = cropBoxData.left;
+            canvasData.maxTop = cropBoxData.top;
+
+            if (viewMode === 2) {
+              if (canvasData.width >= containerWidth) {
+                canvasData.minLeft = min(0, newCanvasLeft);
+                canvasData.maxLeft = max(0, newCanvasLeft);
+              }
+
+              if (canvasData.height >= containerHeight) {
+                canvasData.minTop = min(0, newCanvasTop);
+                canvasData.maxTop = max(0, newCanvasTop);
+              }
+            }
+          }
         } else {
           canvasData.minLeft = -canvasData.width;
           canvasData.minTop = -canvasData.height;
@@ -143,47 +185,46 @@
       }
     },
 
-    renderCanvas: function (changed) {
-      var options = this.options;
+    renderCanvas: function (isChanged) {
       var canvasData = this.canvasData;
       var imageData = this.imageData;
       var rotate = imageData.rotate;
       var naturalWidth = imageData.naturalWidth;
       var naturalHeight = imageData.naturalHeight;
       var aspectRatio;
-      var rotated;
+      var rotatedData;
 
-      if (this.rotated) {
-        this.rotated = false;
+      if (this.isRotated) {
+        this.isRotated = false;
 
         // Computes rotated sizes with image sizes
-        rotated = getRotatedSizes({
+        rotatedData = getRotatedSizes({
           width: imageData.width,
           height: imageData.height,
           degree: rotate
         });
 
-        aspectRatio = rotated.width / rotated.height;
+        aspectRatio = rotatedData.width / rotatedData.height;
 
         if (aspectRatio !== canvasData.aspectRatio) {
-          canvasData.left -= (rotated.width - canvasData.width) / 2;
-          canvasData.top -= (rotated.height - canvasData.height) / 2;
-          canvasData.width = rotated.width;
-          canvasData.height = rotated.height;
+          canvasData.left -= (rotatedData.width - canvasData.width) / 2;
+          canvasData.top -= (rotatedData.height - canvasData.height) / 2;
+          canvasData.width = rotatedData.width;
+          canvasData.height = rotatedData.height;
           canvasData.aspectRatio = aspectRatio;
           canvasData.naturalWidth = naturalWidth;
           canvasData.naturalHeight = naturalHeight;
 
           // Computes rotated sizes with natural image sizes
           if (rotate % 180) {
-            rotated = getRotatedSizes({
+            rotatedData = getRotatedSizes({
               width: naturalWidth,
               height: naturalHeight,
               degree: rotate
             });
 
-            canvasData.naturalWidth = rotated.width;
-            canvasData.naturalHeight = rotated.height;
+            canvasData.naturalWidth = rotatedData.width;
+            canvasData.naturalHeight = rotatedData.height;
           }
 
           this.limitCanvas(true, false);
@@ -227,22 +268,22 @@
 
       this.renderImage();
 
-      if (this.cropped && options.strict) {
+      if (this.isCropped && this.isLimited) {
         this.limitCropBox(true, true);
       }
 
-      if (changed) {
+      if (isChanged) {
         this.output();
       }
     },
 
-    renderImage: function (changed) {
+    renderImage: function (isChanged) {
       var canvasData = this.canvasData;
       var imageData = this.imageData;
-      var reversed;
+      var reversedData;
 
       if (imageData.rotate) {
-        reversed = getRotatedSizes({
+        reversedData = getRotatedSizes({
           width: canvasData.width,
           height: canvasData.height,
           degree: imageData.rotate,
@@ -250,11 +291,11 @@
         }, true);
       }
 
-      extend(imageData, reversed ? {
-        width: reversed.width,
-        height: reversed.height,
-        left: (canvasData.width - reversed.width) / 2,
-        top: (canvasData.height - reversed.height) / 2
+      extend(imageData, reversedData ? {
+        width: reversedData.width,
+        height: reversedData.height,
+        left: (canvasData.width - reversedData.width) / 2,
+        top: (canvasData.height - reversedData.height) / 2
       } : {
         width: canvasData.width,
         height: canvasData.height,
@@ -267,10 +308,10 @@
         'height:' + imageData.height + 'px;' +
         'margin-left:' + imageData.left + 'px;' +
         'margin-top:' + imageData.top + 'px;' +
-        'transform:' + getTransform(imageData)
+        'transform:' + getTransform(imageData) + ';'
       );
 
-      if (changed) {
+      if (isChanged) {
         this.output();
       }
     },
@@ -321,29 +362,29 @@
       this.initialCropBoxData = extend({}, cropBoxData);
     },
 
-    limitCropBox: function (size, position) {
+    limitCropBox: function (isSizeLimited, isPositionLimited) {
       var options = this.options;
-      var strict = options.strict;
+      var aspectRatio = options.aspectRatio;
       var containerData = this.containerData;
       var containerWidth = containerData.width;
       var containerHeight = containerData.height;
       var canvasData = this.canvasData;
       var cropBoxData = this.cropBoxData;
-      var aspectRatio = options.aspectRatio;
+      var isLimited = this.isLimited;
       var minCropBoxWidth;
       var minCropBoxHeight;
       var maxCropBoxWidth;
       var maxCropBoxHeight;
 
-      if (size) {
+      if (isSizeLimited) {
         minCropBoxWidth = num(options.minCropBoxWidth) || 0;
         minCropBoxHeight = num(options.minCropBoxHeight) || 0;
 
         // The min/maxCropBoxWidth/Height must be less than containerWidth/Height
         minCropBoxWidth = min(minCropBoxWidth, containerWidth);
         minCropBoxHeight = min(minCropBoxHeight, containerHeight);
-        maxCropBoxWidth = min(containerWidth, strict ? canvasData.width : containerWidth);
-        maxCropBoxHeight = min(containerHeight, strict ? canvasData.height : containerHeight);
+        maxCropBoxWidth = min(containerWidth, isLimited ? canvasData.width : containerWidth);
+        maxCropBoxHeight = min(containerHeight, isLimited ? canvasData.height : containerHeight);
 
         if (aspectRatio) {
           if (minCropBoxWidth && minCropBoxHeight) {
@@ -372,8 +413,8 @@
         cropBoxData.maxHeight = maxCropBoxHeight;
       }
 
-      if (position) {
-        if (strict) {
+      if (isPositionLimited) {
+        if (isLimited) {
           cropBoxData.minLeft = max(0, canvasData.left);
           cropBoxData.minTop = max(0, canvasData.top);
           cropBoxData.maxLeft = min(containerWidth, canvasData.left + canvasData.width) - cropBoxData.width;
@@ -435,11 +476,11 @@
         'top:' + cropBoxData.top + 'px;'
       );
 
-      if (this.cropped && options.strict) {
+      if (this.isCropped && this.isLimited) {
         this.limitCanvas(true, true);
       }
 
-      if (!this.disabled) {
+      if (!this.isDisabled) {
         this.output();
       }
     },
@@ -449,7 +490,7 @@
 
       this.preview();
 
-      if (this.complete && isFunction(options.crop)) {
+      if (this.isCompleted && isFunction(options.crop)) {
         options.crop.call(this.element, this.getData());
       }
     }