+++ test/options/checkCrossOrigin.js
@@ -2,7 +2,7 @@ window.addEventListener('DOMContentLoaded', function () {
 
   'use strict';
 
-  var crossOriginImage = 'http://fengyuanchen.github.io/cropper/img/picture.jpg';
+  var crossOriginImage = 'https://fengyuanchen.github.io/cropper/img/picture.jpg';
   var image = window.createCropperImage({
         src: crossOriginImage
       });
@@ -11,7 +11,7 @@ window.addEventListener('DOMContentLoaded', function () {
     built: function () {
       var cropper = this.cropper;
 
-      QUnit.test('options.checkImageOrigin', function (assert) {
+      QUnit.test('options.checkCrossOrigin', function (assert) {
         assert.ok(cropper.image.crossOrigin === 'anonymous');
         assert.ok(cropper.image.src.indexOf('timestamp') !== -1);
       });
@@ -29,7 +29,7 @@ window.addEventListener('DOMContentLoaded', function () {
       built: function () {
         var cropper = this.cropper;
 
-        QUnit.test('options.checkImageOrigin: exists crossOrigin attribute', function (assert) {
+        QUnit.test('options.checkCrossOrigin: exists crossOrigin attribute', function (assert) {
           assert.ok(cropper.image.crossOrigin === 'anonymous');
           assert.ok(cropper.image.src.indexOf('timestamp') === -1);
         });